+++ include/interval.cuh
@@ -7,13 +7,13 @@
 
 class Interval {
 public:
-	double min,max;
+	float min,max;
 
 	__host__ __device__ Interval() : min(+INFINITY), max(-INFINITY) {}
 
-	__host__ __device__ Interval(double min,double max): min(min), max(max) {}
+	__host__ __device__ Interval(float min,float max): min(min), max(max) {}
 
-	__host__ __device__ double size() const {return max - min;}
+	__host__ __device__ float size() const {return max - min;}
 
 	/**
 	 * @brief Returns if the value is within, border value returns true.
@@ -21,7 +21,7 @@ public:
 	 * @return true 
 	 * @return false 
 	 */
-	__host__ __device__ bool contains(double x) const {return min <= x && x <= max;}
+	__host__ __device__ bool contains(float x) const {return min <= x && x <= max;}
 
 	/**
 	 * @brief Returns if the value is within, border value returns false.
@@ -29,7 +29,7 @@ public:
 	 * @return true 
 	 * @return false 
 	 */
-	__host__ __device__ bool surrounds(double x) const {return min <= x && x <= max;}
+	__host__ __device__ bool surrounds(float x) const {return min <= x && x <= max;}
 
 	/**
 	 * @brief Expand the interval by delta on both ends.
@@ -38,7 +38,7 @@ public:
 	 * @details This function is used for the bounding box creation for
 	 *          triangles
 	 */
-	__host__ __device__ Interval expand(double delta) const {
+	__host__ __device__ Interval expand(float delta) const {
 		return Interval(min - delta, max + delta);
 	}
 };
